any user can update his own data. only admin can add persons
